[CORE] Js: tests refactor
