+++ core/src/js/jsfuck.rs
@@ -47,7 +47,7 @@ pub struct JsFuckLevelNine;
 
 impl JsFuckLevelNine {
     /// decodes a `Function` body of the form `return '<string literal>'`.
-    fn decode_return_string(body: &str) -> Option<String> {
+    pub(crate) fn decode_return_string(body: &str) -> Option<String> {
         let rest = body.trim().strip_prefix("return")?.trim_start();
         let mut chars = rest.chars();
         let quote = chars.next()?;
@@ -262,44 +262,5 @@ impl<'a> RuleMut<'a> for JsFuckLevelNine {
         }
 
         Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn decode_bmp_escape() {
-        assert_eq!(
-            JsFuckLevelNine::decode_return_string(r"return 'A'").as_deref(),
-            Some("A")
-        );
-    }
-
-    #[test]
-    fn decode_surrogate_pair() {
-        assert_eq!(
-            JsFuckLevelNine::decode_return_string(r"return '😀'").as_deref(),
-            Some("😀")
-        );
-    }
-
-    #[test]
-    fn decode_code_point_and_hex() {
-        assert_eq!(
-            JsFuckLevelNine::decode_return_string(r"return '\u{1F600}'").as_deref(),
-            Some("😀")
-        );
-        assert_eq!(
-            JsFuckLevelNine::decode_return_string(r#"return "\x41\x42""#).as_deref(),
-            Some("AB")
-        );
-    }
-
-    #[test]
-    fn rejects_non_string_return() {
-        assert_eq!(JsFuckLevelNine::decode_return_string("return 1+1"), None);
-        assert_eq!(JsFuckLevelNine::decode_return_string("alert(1)"), None);
     }
 }
