+++ core/src/js/tests/fncall_tests.rs
@@ -1,0 +1,173 @@
+#[cfg(test)]
+mod test_fncall {
+    use crate::js::build_javascript_tree;
+    use crate::js::forward::Forward;
+    use crate::js::functions::fncall::FnCall;
+    use crate::js::functions::function::ParseFunction;
+    use crate::js::integer::{AddInt, ParseInt};
+    use crate::js::linter::Linter;
+    use crate::js::objects::object::{ObjectField, ParseObject};
+    use crate::js::strategy::JavaScriptStrategy;
+    use crate::js::string::ParseString;
+    use crate::js::var::Var;
+
+    fn deobfuscate(input: &str) -> String {
+        let mut tree = build_javascript_tree(input).unwrap();
+        tree.apply_mut_with_strategy(
+            &mut (
+                ParseInt::default(),
+                AddInt::default(),
+                ParseString::default(),
+                ParseFunction::default(),
+                ParseObject::default(),
+                Forward::default(),
+                ObjectField::default(),
+                Var::default(),
+                FnCall::default(),
+            ),
+            JavaScriptStrategy::default(),
+        )
+        .unwrap();
+
+        let mut linter = Linter::default();
+        tree.apply(&mut linter).unwrap();
+        linter.output
+    }
+
+    #[test]
+    fn test_fncall_simple_string_return() {
+        assert_eq!(
+            deobfuscate("function test() { return 'hello'; } console.log(test());"),
+            "function test() { return 'hello'; } console.log('hello');"
+        );
+    }
+
+    #[test]
+    fn test_fncall_fontcolor_with_arg() {
+        assert_eq!(
+            deobfuscate("console.log('minusone'.fontcolor('red'));"),
+            "console.log('<font color=\"red\">minusone</font>');"
+        );
+    }
+
+    #[test]
+    fn test_fncall_fontcolor_escapes_quote_in_arg() {
+        assert_eq!(
+            deobfuscate("console.log(''.fontcolor('0false\"'));"),
+            "console.log('<font color=\"0false&quot;\"></font>');"
+        );
+    }
+
+    #[test]
+    fn test_fncall_simple_int_return() {
+        assert_eq!(
+            deobfuscate("function getValue() { return 42; } var x = getValue();"),
+            "function getValue() { return 42; } var x = 42;"
+        );
+    }
+
+    #[test]
+    fn test_fncall_with_var_inside() {
+        assert_eq!(
+            deobfuscate("function test() { var a = 'hello'; return a; } console.log(test());"),
+            "function test() { var a = 'hello'; return 'hello'; } console.log('hello');"
+        );
+    }
+
+    #[test]
+    fn test_fncall_does_not_resolve_param_dependent_return() {
+        assert_eq!(
+            deobfuscate("function test(x) { return x; } console.log(test('hello'));"),
+            "function test(x) { return x; } console.log(test('hello'));"
+        );
+    }
+
+    #[test]
+    fn test_fncall_resolve_param_independent_return() {
+        assert_eq!(
+            deobfuscate("function test(x) { console.log(x); return 1; } var a = test(7);"),
+            "function test(x) { console.log(x); return 1; } var a = 1;"
+        );
+    }
+
+    #[test]
+    fn test_fncall_resolve_with_args_when_return_is_constant() {
+        assert_eq!(
+            deobfuscate("function test() { return 'hello'; } console.log(test('unused'));"),
+            "function test() { return 'hello'; } console.log('hello');"
+        );
+    }
+
+    #[test]
+    fn test_fncall_multiple_returns_not_resolved() {
+        assert_eq!(
+            deobfuscate(
+                "function test() { if (true) { return 'a'; } return 'b'; } console.log(test());"
+            ),
+            "function test() { if (true) { return 'a'; } return 'b'; } console.log(test());"
+        );
+    }
+
+    #[test]
+    fn test_fncall_no_return_not_resolved() {
+        assert_eq!(
+            deobfuscate("function test() { var a = 1; } console.log(test());"),
+            "function test() { var a = 1; } console.log(test());"
+        );
+    }
+
+    #[test]
+    fn test_fncall_nested_function_scope() {
+        assert_eq!(
+            deobfuscate(
+                "function outer() { function inner() { return 'inner'; } return 'outer'; } console.log(outer());"
+            ),
+            "function outer() { function inner() { return 'inner'; } return 'outer'; } console.log('outer');"
+        );
+    }
+
+    #[test]
+    fn test_fncall_expression_return() {
+        assert_eq!(
+            deobfuscate("function test() { return 1 + 2; } console.log(test());"),
+            "function test() { return 3; } console.log(3);"
+        );
+    }
+
+    #[test]
+    fn test_fncall_unknown_return_not_resolved() {
+        assert_eq!(
+            deobfuscate("function test() { return foo(); } console.log(test());"),
+            "function test() { return foo(); } console.log(test());"
+        );
+    }
+
+    #[test]
+    fn test_fncall_object_stored_function_constant_return() {
+        assert_eq!(
+            deobfuscate(
+                "let a = {}; let x = function (params) { return 0; } a.t = x; console.log(a.t());"
+            ),
+            "let a = {}; let x = function (params) { return 0; } a.t = x; console.log(0);"
+        );
+    }
+
+    #[test]
+    fn test_fncall_object_stored_function_param_dependent_return() {
+        assert_eq!(
+            deobfuscate(
+                "let a = {}; let x = function (n) { return n+1; } a.t = x; console.log(a.t(1)); console.log(a.t(2));"
+            ),
+            "let a = {}; let x = function (n) { return n+1; } a.t = x; console.log(2); console.log(3);"
+        );
+    }
+
+    #[test]
+    fn test_fncall_self_redefining_function() {
+        let output = deobfuscate(
+            "function _0x45a5(){return(_0x45a5=function(){return'minusone'})()}console.log(_0x45a5());",
+        );
+
+        assert!(output.ends_with("console.log('minusone');"));
+    }
+}
