+++ core/src/js/tests/mod.rs
@@ -1,11 +1,20 @@
 mod array_tests;
 mod b64_tests;
 mod bool_tests;
+mod buffer_tests;
 mod comparator_tests;
+mod converter_tests;
 mod encode_decode_tests;
+mod fncall_tests;
+mod forward_tests;
+mod function_tests;
 mod integer_tests;
-mod jsfuck;
+mod jsfuck_tests;
+mod linter_tests;
+mod loop_tests;
 mod maths_tests;
+mod object_tests;
+mod objectify_tests;
 mod post_process_tests;
 mod regex_tests;
 mod specials_tests;
